requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.11.9
pydantic-settings==2.11.0
langchain-anthropic==0.3.20
langchain-community==0.3.0
anthropic==0.68.1
sqlalchemy==2.0.23
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2